Cleanup some compiler warnings seen with clang on OS X.

+++ pgadmin/ctl/ctlSQLGrid.cpp
@@ -348,7 +348,7 @@ void ctlSQLGrid::OnLabelDoubleClick(wxGridEvent &event)
 		else // toggle between some predefined sizes
 		{
 
-			if (col < colMaxSizes.GetCount() && colMaxSizes[col] >= 0)
+			if (col < (int)colMaxSizes.GetCount() && colMaxSizes[col] >= 0)
 				extentWant = colMaxSizes[col];
 			else
 			{
@@ -395,7 +395,7 @@ void ctlSQLGrid::OnLabelDoubleClick(wxGridEvent &event)
 
 void ctlSQLGrid::AutoSizeColumn(int col, bool setAsMin, bool doLimit)
 {
-	if (col < colMaxSizes.GetCount() && colMaxSizes[col] >= 0)
+	if (col < (int)colMaxSizes.GetCount() && colMaxSizes[col] >= 0)
 		SetColSize(col, colMaxSizes[col]);
 	else
 		wxGrid::AutoSizeColumn(col, setAsMin);
@@ -490,7 +490,7 @@ void ctlSQLGrid::AutoSizeColumns(bool setAsMin)
 		{
 			oldSize = GetColSize(col);
 			// Is too wide and no user-specified size
-			if (oldSize > maxSize && !(col < colMaxSizes.GetCount() && colMaxSizes[col] == -1))
+			if (oldSize > maxSize && !(col < (int)colMaxSizes.GetCount() && colMaxSizes[col] == -1))
 			{
 				totalSize -= oldSize;
 				/* Shrink wide column to maxSize.
